Committing working homepage with bootstrap example page
I've set up the necessary angularjs controllers and services to display
a test page that I got from an example from bootstrapjs. I've also set
up a service that will build the Yelp API query from the authentication
keys and whatever parameters are passed in. But due to cross domain
issues we're going to have to use a proxy to make the API queries so I
stole the one given to use from assignment 1. I'll get the API querying
working tomorrow.